fixes to montgomery's code
montgomery asked for a function that takes in the choice of door and location of the car, not to assign them within the function. So I got rid of the sampling within the function and made a separate function first to randomly select the placement of the car. Because the user gets to choose their own door I let this just be a numerical input into the second function and I got rid of x because it was unneccessary. I also renamed the variables and the function to be more clear. Then I simplified the if then statement to simply return true if the inputs of the choice of door = the placement of the car and to return false otherwise.