bug fixes
realized part of my chooseDoor function wasn't doing what I thought it was - it was selecting 1 every time because I had forced a command that output a logical to be numeric. So instead I created a function to filter a vector of 1:3 and take out the values of the initial chosen door, the car door, and the door to be removed as necessary and then to randomly select 1 value in case there were 2 left over. This fixed the issue.